move more of the state into the shared lib
