+++ common/credential-proxy/src/lib.rs
@@ -4,5 +4,7 @@
 pub mod deposits_buffer;
 pub mod error;
 pub mod helpers;
+pub mod nym_api_helpers;
+pub mod quorum_checker;
 pub mod shared_state;
 pub mod storage;
